Started the time steward rewrite
Specifically, this rewrite does 2 major things:
1) use the predictor model instead of the trigger model
2) use a simpler extended_time type that doesn't require the manual
orderer

This rewrite is partially in preparation for making the time steward
able to serialize snapshots. The trigger model did not provide any
reasonable way to implement serialization. The predictor model does,
since all data at any specific time is contained in the entity states at
that time.

This rewrite also optimizes the storage of field data by using ordered
std::vectors instead of std::sets. There are some downsides to being
unable to delete in the middle, but in most cases this should be much
faster.